feat: let-polymorphism, poly stubs, array type sugar, named closures
Type system
- dep: instantiate global types with fresh metavariables at each call
  site (dep_freshen_type) so (id "ciao") and (id 3) no longer interfere
- infer: bare unannotated params now get fresh type variables instead of
  TYPE_COLL, fixing polymorphic functions like (define double x -> 2*x)
- reader: assign synthetic type variables ('a', 'b', …) to unannotated
  params so the dep checker sees genuine ∀-quantified signatures
- dep: handle (define [name :: Type] value) typed variable bindings

Codegen
- Emit unreachable + ExternalWeakLinkage for fully-polymorphic stubs;
  codegen_specialize produces typed versions at each concrete call site
- Fix type derivation at specialization sites: prefer literal kind and
  env lookup over stale inferred_type, which could be stamped by a prior
  call site
- Fix double-free in codegen_specialize: create child env before any
  insertions, restore correctly through spec_outer_env

Runtime / closures
- Add rt_value_closure_named; closures now carry an optional debug name
- Pretty-print closures as "name/arity" instead of "<closure/N>"
- Switch wrap_func_as_closure to rt_value_closure_named throughout

Array / memory sugar
- [16kb], [16mb I32], [4096 I32] size-literal syntax in bracket lists
- define buf :: [512 U8] and define buf :: [16kb] variable declarations
  desugar to (define [buf :: Arr :: N :: U8] [])
- *U8 pointer-type sugar expands to Pointer :: U8 (reader + wisp)
- Arr :: Size :: ElemType and Arr :: ElemType :: Size both accepted
- Architecture-aware static array size limit with actionable error
- arr_size promoted from int to int64_t throughout

Misc
- pmatch_rename_anon_params: propagate PAT_VAR names from first clause
  back into anonymous params so codegen sees real argument names
- LLVMArrayType → LLVMArrayType2 (avoids deprecated API)
- Compiler directive stubs added to TODO (PURE, CamelCase, ASM modules)

+++ types.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef enum {
     TYPE_INT,
@@ -90,7 +91,7 @@ typedef struct Type {
 
     // TYPE_ARR — fat pointer { T* data, i64 size }
     struct Type *arr_element_type;
-    int          arr_size;          // -1 = unknown at compile time
+    int64_t      arr_size;          // -1 = unknown at compile time
     bool         arr_is_fat;        // true = runtime fat pointer
 
     // TYPE_LAYOUT
@@ -173,7 +174,7 @@ Type *type_int_arbitrary(int width, bool is_signed); // I<n> or U<n>
 /// Constructors — compound types
 
 Type *type_list(Type **types, int count);
-Type *type_arr(Type *element_type, int size);
+Type *type_arr(Type *element_type, int64_t size);
 Type *type_arr_fat(Type *element_type); // runtime fat pointer {data, size}
 Type *type_fn(FnParam *params, int param_count, Type *return_type);
 Type *type_fn_builtin(int min_args, int opt_args, bool variadic);
